requirements.txt
requests>=2.28.0
Pillow>=9.0.0
reportlab>=4.0.0
